Implement shimmer on screens missing it and implement color pills per investment category
