+++ app/src/main/java/com/prometheo/moneylife/ui/investments/current/CurrentInvestmentsFragment.kt
@@ -1,4 +1,4 @@
-package com.prometheo.moneylife.ui.investments
+package com.prometheo.moneylife.ui.investments.current
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -11,7 +11,8 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.prometheo.moneylife.R
-import com.prometheo.moneylife.databinding.FragmentCurrentInvestmentsBinding
+import com.prometheo.moneylife.databinding.FragmentInvestmentsBinding
+import com.prometheo.moneylife.ui.investments.available.AvailableInvestmentsFragment
 import com.xwray.groupie.GroupAdapter
 import com.xwray.groupie.GroupieViewHolder
 import dagger.hilt.android.AndroidEntryPoint
@@ -25,7 +26,7 @@ class CurrentInvestmentsFragment : Fragment() {
 
     private val vm: CurrentInvestmentsViewModel by viewModels()
     private val adapter = GroupAdapter<GroupieViewHolder>()
-    private var _binding: FragmentCurrentInvestmentsBinding? = null
+    private var _binding: FragmentInvestmentsBinding? = null
     private val binding get() = _binding!!
 
     private var currentEditingInvestmentPosition: Int? = null
@@ -35,7 +36,7 @@ class CurrentInvestmentsFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentCurrentInvestmentsBinding.inflate(inflater, container, false).apply {
+        _binding = FragmentInvestmentsBinding.inflate(inflater, container, false).apply {
             rv.adapter = adapter
             rv.layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
 
@@ -50,6 +51,8 @@ class CurrentInvestmentsFragment : Fragment() {
 
         vm.state.observe(viewLifecycleOwner, Observer { state ->
             binding.loadingIndicator.isVisible = state.loading
+            binding.shimmerViewContainer.isVisible = state.loading
+            binding.rv.isVisible = !state.loading
         })
 
         vm.currentInvestments.observe(viewLifecycleOwner, Observer { currentInvestments ->
