+++ app/src/main/res/layout/fragment_investments.xml
@@ -5,7 +5,21 @@
     android:layout_height="match_parent"
     android:background="@color/app_background"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.investments.CurrentInvestmentsFragment">
+    tools:context=".ui.investments.current.CurrentInvestmentsFragment">
+
+    <com.facebook.shimmer.ShimmerFrameLayout
+        android:id="@+id/shimmer_view_container"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:visibility="gone">
+
+        <include layout="@layout/placeholder_investments" />
+
+    </com.facebook.shimmer.ShimmerFrameLayout>
 
     <TextView
         android:id="@+id/title"
