+++ app/src/main/res/layout/item_owned_investment.xml
@@ -119,6 +119,8 @@
                 android:layout_height="wrap_content"
                 android:layout_marginTop="32dp"
                 android:hint="@string/amount_placeholder"
+                android:visibility="gone"
+                tools:visibility="visible"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toBottomOf="@id/initialContributionLabel">
@@ -138,6 +140,8 @@
                 android:layout_marginTop="16dp"
                 android:layout_marginEnd="8dp"
                 android:text="@string/invest_button"
+                android:visibility="gone"
+                tools:visibility="visible"
                 app:backgroundTint="#31B518"
                 app:layout_constraintBottom_toTopOf="@id/withdrawAllButton"
                 app:layout_constraintEnd_toStartOf="@id/withdrawButton"
@@ -154,11 +158,12 @@
                 android:layout_marginTop="16dp"
                 android:text="@string/withdraw_button"
                 android:textColor="#FFFFFF"
+                android:visibility="gone"
+                tools:visibility="visible"
                 app:backgroundTint="#EE0000"
                 app:layout_constraintBottom_toTopOf="@id/withdrawAllButton"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.5"
-
                 app:layout_constraintStart_toEndOf="@id/investButton"
                 app:layout_constraintTop_toBottomOf="@id/amountFieldHint"
                 app:layout_constraintVertical_bias="0.111" />
@@ -171,6 +176,8 @@
                 android:layout_marginTop="8dp"
                 android:text="@string/withdraw_all_button"
                 android:textColor="#EE0000"
+                android:visibility="gone"
+                tools:visibility="visible"
                 app:backgroundTint="#10EE0000"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintStart_toStartOf="parent"
