+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -25,9 +25,9 @@
 
     <fragment
         android:id="@+id/navigation_investments"
-        android:name="com.prometheo.moneylife.ui.investments.CurrentInvestmentsFragment"
+        android:name="com.prometheo.moneylife.ui.investments.current.CurrentInvestmentsFragment"
         android:label="@string/title_investments"
-        tools:layout="@layout/fragment_current_investments" />
+        tools:layout="@layout/fragment_investments" />
 
     <fragment
         android:id="@+id/navigation_notifications"
